feat: Remove old BotControl page and integrate JS into settings (#1275)
