+++ src/DiscordBot.Bot/Pages/Shared/Components/_RestartBanner.cshtml
@@ -11,9 +11,10 @@
                 <h3 class="text-base font-semibold text-warning mb-1">Restart Required</h3>
                 <p class="text-sm text-text-primary mb-3">
                     You have made changes to settings that require a bot restart to take effect.
-                    Please restart the bot from the Bot Control panel to apply these changes.
+                    Please go to the Bot Control tab in Settings to restart the bot.
                 </p>
-                <a href="/Admin/BotControl"
+                <a href="/Admin/Settings"
+                   onclick="window.settingsManager?.switchTab('BotControl'); return true;"
                    class="inline-flex items-center gap-2 px-4 py-2 bg-warning hover:bg-amber-600 active:bg-amber-700 text-text-inverse font-semibold text-sm rounded-lg transition-colors">
                     <svg class="w-4 h-4" fill="none" viewBox="0 0 24 24" stroke="currentColor">
                         <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />
