+++ src/DiscordBot.Bot/Pages/Shared/Components/_TypedConfirmationModal.cshtml
@@ -4,7 +4,7 @@
 <!-- Typed Confirmation Modal -->
 <div id="@Model.Id" class="hidden fixed inset-0 z-50" role="alertdialog" aria-modal="true" aria-labelledby="@(Model.Id)Title" aria-describedby="@(Model.Id)Desc">
     <!-- Backdrop -->
-    <div class="fixed inset-0 bg-black/70 backdrop-blur-sm" onclick="window.botControl?.hideTypedModal('@Model.Id')"></div>
+    <div class="fixed inset-0 bg-black/70 backdrop-blur-sm" onclick="window.settingsManager?.hideTypedModal('@Model.Id')"></div>
 
     <!-- Modal Dialog -->
     <div class="fixed inset-0 flex items-center justify-center p-4">
@@ -37,7 +37,7 @@
                                 autocomplete="off"
                                 data-required-text="@Model.RequiredText"
                                 data-confirm-btn="@(Model.Id)ConfirmBtn"
-                                oninput="window.botControl?.validateTypedInput(this)"
+                                oninput="window.settingsManager?.validateTypedInput(this)"
                             />
                         </div>
                     </div>
@@ -47,7 +47,7 @@
             <!-- Modal Footer -->
             <div class="flex justify-end gap-3 px-6 py-4 bg-bg-secondary border-t border-border-primary rounded-b-lg">
                 <button type="button"
-                        onclick="window.botControl?.hideTypedModal('@Model.Id')"
+                        onclick="window.settingsManager?.hideTypedModal('@Model.Id')"
                         class="px-4 py-2 bg-bg-tertiary hover:bg-bg-hover border border-border-primary text-text-primary font-medium text-sm rounded-md transition-colors">
                     @Model.CancelText
                 </button>
