+++ src/DiscordBot.Bot/Pages/Shared/_Sidebar.cshtml
@@ -107,16 +107,6 @@
           <span class="sidebar-link-text">Users</span>
         </a>
         @{
-          var botControlActive = ViewContext.RouteData.Values["page"]?.ToString() == "/Admin/BotControl";
-        }
-        <a href="/Admin/BotControl" class="sidebar-link-redesign @(botControlActive ? "active" : "")" @(botControlActive ? "aria-current=page" : "")>
-          <svg class="w-5 h-5 flex-shrink-0" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z" />
-            <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
-          </svg>
-          <span class="sidebar-link-text">Bot Control</span>
-        </a>
-        @{
           var messageLogsActive = ViewContext.RouteData.Values["page"]?.ToString()?.StartsWith("/Admin/MessageLogs") == true;
         }
         <a href="/Admin/MessageLogs" class="sidebar-link-redesign @(messageLogsActive ? "active" : "")" @(messageLogsActive ? "aria-current=page" : "")>
